reading xml
so I have managed to get it to read the xml docs, but there was some
problems where the program would crash when facing an unknown
character.

I have tried to fix this using codecs module. But this hasn’t fixed the
problem yet.

committing before I make more changes